breakout process defs and templates
