+++ pkg/codegen/operations.go
@@ -18,13 +18,11 @@ import (
 	"bytes"
 	"fmt"
 	"sort"
-	"strconv"
 	"strings"
 	"text/template"
 
 	"github.com/getkin/kin-openapi/openapi3"
 
-	"github.com/oapi-codegen/oapi-codegen/v2/pkg/codegen/openapiv3"
 	"github.com/oapi-codegen/oapi-codegen/v2/pkg/util"
 )
 
@@ -77,162 +75,6 @@ func DescribeSecurityDefinition(securityRequirements openapi3.SecurityRequiremen
 	}
 
 	return outDefs
-}
-
-// RequestBodyDefinition describes a request body
-type RequestBodyDefinition struct {
-	// Is this body required, or optional?
-	Required bool
-
-	// This is the schema describing this body
-	Schema Schema
-
-	// When we generate type names, we need a Tag for it, such as JSON, in
-	// which case we will produce "JSONBody".
-	NameTag string
-
-	// This is the content type corresponding to the body, eg, application/json
-	ContentType string
-
-	// Whether this is the default body type. For an operation named OpFoo, we
-	// will not add suffixes like OpFooJSONBody for this one.
-	Default bool
-
-	// Contains encoding options for formdata
-	Encoding map[string]RequestBodyEncoding
-}
-
-// TypeDef returns the Go type definition for a request body
-func (r RequestBodyDefinition) TypeDef(opID string) *TypeDefinition {
-	return &TypeDefinition{
-		TypeName: fmt.Sprintf("%s%sRequestBody", opID, r.NameTag),
-		Schema:   r.Schema,
-	}
-}
-
-// CustomType returns whether the body is a custom inline type, or pre-defined. This is
-// poorly named, but it's here for compatibility reasons post-refactoring
-// TODO: clean up the templates code, it can be simpler.
-func (r RequestBodyDefinition) CustomType() bool {
-	return r.Schema.RefType == ""
-}
-
-// When we're generating multiple functions which relate to request bodies,
-// this generates the suffix. Such as Operation DoFoo would be suffixed with
-// DoFooWithXMLBody.
-func (r RequestBodyDefinition) Suffix() string {
-	// The default response is never suffixed.
-	if r.Default {
-		return ""
-	}
-	return "With" + r.NameTag + "Body"
-}
-
-// IsSupportedByClient returns true if we support this content type for client. Otherwise only generic method will ge generated
-func (r RequestBodyDefinition) IsSupportedByClient() bool {
-	return r.IsJSON() || r.NameTag == "Formdata" || r.NameTag == "Text"
-}
-
-// IsJSON returns whether this is a JSON media type, for instance:
-// - application/json
-// - application/vnd.api+json
-// - application/*+json
-func (r RequestBodyDefinition) IsJSON() bool {
-	return util.IsMediaTypeJson(r.ContentType)
-}
-
-// IsSupported returns true if we support this content type for server. Otherwise io.Reader will be generated
-func (r RequestBodyDefinition) IsSupported() bool {
-	return r.NameTag != ""
-}
-
-// IsFixedContentType returns true if content type has fixed content type, i.e. contains no "*" symbol
-func (r RequestBodyDefinition) IsFixedContentType() bool {
-	return !strings.Contains(r.ContentType, "*")
-}
-
-type RequestBodyEncoding struct {
-	ContentType string
-	Style       string
-	Explode     *bool
-}
-
-type ResponseDefinition struct {
-	StatusCode  string
-	Description string
-	Contents    []ResponseContentDefinition
-	Headers     []ResponseHeaderDefinition
-	Ref         string
-}
-
-func (r ResponseDefinition) HasFixedStatusCode() bool {
-	_, err := strconv.Atoi(r.StatusCode)
-	return err == nil
-}
-
-func (r ResponseDefinition) GoName() string {
-	return SchemaNameToTypeName(r.StatusCode)
-}
-
-func (r ResponseDefinition) IsRef() bool {
-	return r.Ref != ""
-}
-
-func (r ResponseDefinition) IsExternalRef() bool {
-	if !r.IsRef() {
-		return false
-	}
-	return strings.Contains(r.Ref, ".")
-}
-
-type ResponseContentDefinition struct {
-	// This is the schema describing this content
-	Schema Schema
-
-	// This is the content type corresponding to the body, eg, application/json
-	ContentType string
-
-	// When we generate type names, we need a Tag for it, such as JSON, in
-	// which case we will produce "Response200JSONContent".
-	NameTag string
-}
-
-// TypeDef returns the Go type definition for a request body
-func (r ResponseContentDefinition) TypeDef(opID string, statusCode int) *TypeDefinition {
-	return &TypeDefinition{
-		TypeName: fmt.Sprintf("%s%v%sResponse", opID, statusCode, r.NameTagOrContentType()),
-		Schema:   r.Schema,
-	}
-}
-
-func (r ResponseContentDefinition) IsSupported() bool {
-	return r.NameTag != ""
-}
-
-// HasFixedContentType returns true if content type has fixed content type, i.e. contains no "*" symbol
-func (r ResponseContentDefinition) HasFixedContentType() bool {
-	return !strings.Contains(r.ContentType, "*")
-}
-
-func (r ResponseContentDefinition) NameTagOrContentType() string {
-	if r.NameTag != "" {
-		return r.NameTag
-	}
-	return SchemaNameToTypeName(r.ContentType)
-}
-
-// IsJSON returns whether this is a JSON media type, for instance:
-// - application/json
-// - application/vnd.api+json
-// - application/*+json
-func (r ResponseContentDefinition) IsJSON() bool {
-	return util.IsMediaTypeJson(r.ContentType)
-}
-
-type ResponseHeaderDefinition struct {
-	Name   string
-	GoName string
-	Schema Schema
 }
 
 // FilterParameterDefinitionByType returns the subset of the specified parameters which are of the
@@ -683,106 +525,6 @@ func GenerateTypesForOperations(t *template.Template, ops []OperationDefinition)
 	}
 
 	return buf.String(), nil
-}
-
-// GenerateIrisServer generates all the go code for the ServerInterface as well as
-// all the wrapper functions around our handlers.
-func GenerateIrisServer(t *template.Template, operations []OperationDefinition) (string, error) {
-	return GenerateTemplates([]string{"iris/iris-interface.tmpl", "iris/iris-middleware.tmpl", "iris/iris-handler.tmpl"}, t, operations)
-}
-
-// GenerateChiServer generates all the go code for the ServerInterface as well as
-// all the wrapper functions around our handlers.
-func GenerateChiServer(t *template.Template, operations []OperationDefinition) (string, error) {
-	return GenerateTemplates([]string{"chi/chi-interface.tmpl", "chi/chi-middleware.tmpl", "chi/chi-handler.tmpl"}, t, operations)
-}
-
-// GenerateFiberServer generates all the go code for the ServerInterface as well as
-// all the wrapper functions around our handlers.
-func GenerateFiberServer(t *template.Template, operations []OperationDefinition) (string, error) {
-	return GenerateTemplates([]string{"fiber/fiber-interface.tmpl", "fiber/fiber-middleware.tmpl", "fiber/fiber-handler.tmpl"}, t, operations)
-}
-
-// GenerateEchoServer generates all the go code for the ServerInterface as well as
-// all the wrapper functions around our handlers.
-func GenerateEchoServer(t *template.Template, operations []OperationDefinition) (string, error) {
-	return GenerateTemplates([]string{"echo/echo-interface.tmpl", "echo/echo-wrappers.tmpl", "echo/echo-register.tmpl"}, t, operations)
-}
-
-// GenerateGinServer generates all the go code for the ServerInterface as well as
-// all the wrapper functions around our handlers.
-func GenerateGinServer(t *template.Template, operations []OperationDefinition) (string, error) {
-	return GenerateTemplates([]string{"gin/gin-interface.tmpl", "gin/gin-wrappers.tmpl", "gin/gin-register.tmpl"}, t, operations)
-}
-
-// GenerateGorillaServer generates all the go code for the ServerInterface as well as
-// all the wrapper functions around our handlers.
-func GenerateGorillaServer(t *template.Template, operations []OperationDefinition) (string, error) {
-	return GenerateTemplates([]string{"gorilla/gorilla-interface.tmpl", "gorilla/gorilla-middleware.tmpl", "gorilla/gorilla-register.tmpl"}, t, operations)
-}
-
-// GenerateStdHTTPServer generates all the go code for the ServerInterface as well as
-// all the wrapper functions around our handlers.
-func GenerateStdHTTPServer(t *template.Template, operations []OperationDefinition) (string, error) {
-	return GenerateTemplates([]string{"stdhttp/std-http-interface.tmpl", "stdhttp/std-http-middleware.tmpl", "stdhttp/std-http-handler.tmpl"}, t, operations)
-}
-
-func GenerateStrictServer(t *template.Template, operations []OperationDefinition, opts openapiv3.Configuration) (string, error) {
-
-	var templates []string
-
-	if opts.Generate.ChiServer || opts.Generate.GorillaServer || opts.Generate.StdHTTPServer {
-		templates = append(templates, "strict/strict-interface.tmpl", "strict/strict-http.tmpl")
-	}
-	if opts.Generate.EchoServer {
-		templates = append(templates, "strict/strict-interface.tmpl", "strict/strict-echo.tmpl")
-	}
-	if opts.Generate.GinServer {
-		templates = append(templates, "strict/strict-interface.tmpl", "strict/strict-gin.tmpl")
-	}
-	if opts.Generate.FiberServer {
-		templates = append(templates, "strict/strict-fiber-interface.tmpl", "strict/strict-fiber.tmpl")
-	}
-	if opts.Generate.IrisServer {
-		templates = append(templates, "strict/strict-iris-interface.tmpl", "strict/strict-iris.tmpl")
-	}
-
-	return GenerateTemplates(templates, t, operations)
-}
-
-func GenerateStrictResponses(t *template.Template, responses []ResponseDefinition) (string, error) {
-	return GenerateTemplates([]string{"strict/strict-responses.tmpl"}, t, responses)
-}
-
-// GenerateClient uses the template engine to generate the function which registers our wrappers
-// as Echo path handlers.
-func GenerateClient(t *template.Template, ops []OperationDefinition) (string, error) {
-	return GenerateTemplates([]string{"client.tmpl"}, t, ops)
-}
-
-// GenerateClientWithResponses generates a client which extends the basic client which does response
-// unmarshaling.
-func GenerateClientWithResponses(t *template.Template, ops []OperationDefinition) (string, error) {
-	return GenerateTemplates([]string{"client-with-responses.tmpl"}, t, ops)
-}
-
-// GenerateTemplates used to generate templates
-func GenerateTemplates(templates []string, t *template.Template, ops interface{}) (string, error) {
-	var generatedTemplates []string
-	for _, tmpl := range templates {
-		var buf bytes.Buffer
-		w := bufio.NewWriter(&buf)
-
-		if err := t.ExecuteTemplate(w, tmpl, ops); err != nil {
-			return "", fmt.Errorf("error generating %s: %s", tmpl, err)
-		}
-		if err := w.Flush(); err != nil {
-			return "", fmt.Errorf("error flushing output buffer for %s: %s", tmpl, err)
-		}
-		generatedTemplates = append(generatedTemplates, buf.String())
-	}
-
-	return strings.Join(generatedTemplates, "\n"), nil
 }
 
 // CombineOperationParameters combines the Parameters defined at a global level (Parameters defined for all methods on a given path) with the Parameters defined at a local level (Parameters defined for a specific path), preferring the locally defined parameter over the global one
